Audit src/apps: Fix 64-bit integer handling in Icon-O-Matic
- **NummericalTextView.h/cpp**: Refactored to support `int64` values. Replaced `int32 fIntValueCache` with `int64 fInt64ValueCache`. Added `SetValue(int64)` and `Int64Value()` methods using `atoll` for parsing. Updated `IntValue()` to cast from `int64`. Updated `Invoke` to include the 64-bit value in the message.
- **Int64ValueView.cpp**: Updated to use the new `Int64Value()` and `SetValue(int64)` methods of `NummericalTextView`, removing the previous `int32` casts and resolving associated TODOs.

This change ensures that 64-bit integer properties in Icon-O-Matic can be edited correctly without truncation to 32-bit.

+++ src/apps/icon-o-matic/generic/property/view/specific_properties/Int64ValueView.cpp
@@ -25,9 +25,8 @@ Int64ValueView::Int64ValueView(Int64Property* property)
 	AddChild(fTextView);
 	fTextView->SetFloatMode(false);
 
-// TODO: make NummericalTextView support int64?
 	if (fProperty)
-		fTextView->SetValue((int32)fProperty->Value());
+		fTextView->SetValue(fProperty->Value());
 }
 
 // destructor
@@ -47,9 +46,8 @@ void
 Int64ValueView::ValueChanged()
 {
 	if (fProperty) {
-		fProperty->SetValue(fTextView->IntValue());
-// TODO: make NummericalTextView support int64?
-		fTextView->SetValue((int32)fProperty->Value());
+		fProperty->SetValue(fTextView->Int64Value());
+		fTextView->SetValue(fProperty->Value());
 		TextInputValueView::ValueChanged();
 	}
 }
@@ -60,9 +58,8 @@ Int64ValueView::AdoptProperty(Property* property)
 {
 	Int64Property* p = dynamic_cast<Int64Property*>(property);
 	if (p) {
-// TODO: make NummericalTextView support int64?
-		if (fTextView->IntValue() != (int32)p->Value())
-			fTextView->SetValue((int32)p->Value());
+		if (fTextView->Int64Value() != p->Value())
+			fTextView->SetValue(p->Value());
 		fProperty = p;
 		return true;
 	}
